Layout fixes + empty state
- Fix the icons.
- Add an empty state.
- Improve the layout of items.
- Move ActivityItem in its own component.
