+++ src/components/Activity/IconEmptyState.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+const IconEmptyState = props => (
+  <svg width={30} height={30} fill="none" {...props}>
+    <path
+      d="M10.8 4.7v20.6l-4.7-4.7M19.2 25.3V4.7l4.7 4.7"
+      stroke="#B3B3B3"
+      strokeWidth={1.4}
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+)
+
+export default IconEmptyState
