+++ src/icons/IconPending.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+const IconPending = props => (
+  <svg width={22} height={22} fill="none" {...props}>
+    <rect x={4} y={4} width={14} height={14} rx={7} fill="#979797" />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M10.16 12.547h1.298V7.3H10.16v5.247zM10.05 15h1.529v-1.463h-1.53V15z"
+      fill="#fff"
+    />
+  </svg>
+)
+
+export default IconPending
