+++ src/icons/IconSuccess.js
@@ -1,0 +1,20 @@
+import React from 'react'
+
+const IconSuccess = props => (
+  <svg width={22} height={22} fill="none" {...props}>
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M11 18a7 7 0 1 0 0-14 7 7 0 0 0 0 14z"
+      fill="#21D48E"
+    />
+    <path
+      d="M8 10.523l1.815 2.213L14 9"
+      stroke="#fff"
+      strokeWidth={1.5}
+      strokeLinecap="square"
+    />
+  </svg>
+)
+
+export default IconSuccess
